requirements.txt
Flask==3.0.3
flask-cors==4.0.1
faster-whisper==1.0.3
tqdm==4.66.5
